[Release] Support the --remote option in bump-oss-version.js script
In my RN checkout, I use "upstream" as my remote instead of "origin" -> this lets me run `scripts/bump-oss-version.js --remote upstream 0.41.1` for example.

Also made the script executable so we don't need to put `node` in front of it, and updated the Releases.md doc.

Test Plan: Did a dry run, saw that without the `--remote` flag the default remote is still `origin` and that specifying the flag let me override the origin.