Potential Re-use of delegate React Native code
I'm currently not to sold on the re-use of this commit, but wanted to
take it a step forward and see if anything good could come out of it. I
welcome strict critticism and hope to improve this code. I still think a
Fragment would be valuable, but sharing the necessary methods between
the ReactActivity and making sure to point out methods that can't be
shared is required.