Create buck module modules/debug:interfaces
The purpose of this module is that later we're going to create a devsupport:interfaces module. This module needs to be able to depend on DeveloperSettings without depending on all of devsupport. If it dependen on all of devsupport, then we'd have this cycle:
  uimanager -> modules/core -> devsupport:interfaces -> modules/debug -> uimanager

New module contains:
  - DeveloperSettings

These modules depend on it:
  - modules/debug
  - devsupport
  - react