Remove modules/core's dependency on devsupport
This change addresses the following cycle:
  uimanager -> modules/core -> devsupport -> modules/debug -> uimanager

We break this cycle by removing modules/core's dependency on devsupport. Instead, it depends on devsupport:interfaces.

Thus, the dependency chain becomes:
  -   uimanager -> modules/core -> devsupport:interfaces -> modules/debug:interfaces

New module contains:
  - DevOptionHandler
  - DevSupportManager
  - PackagerStatusCallback
  - StackFrame

The PackagerStatusCallback interface was extracted from the class DevServerHelper and moved into its own file. PackagerStatusCallback is in devsupport:interfaces and DevServerHelper remains in devsupport.

StackFrame is a new interface. There used to be a class called StackFrame but it has been renamed to StackFrameImpl and it implements StackFrame. StackFrame is in devsupport:interfaces and StackFrameImpl is in devsupport.

These modules depend on the new module:
  - devsupport
  - modules/core
  - react