Allow Windows to use assets unbundle
In `react-native-windows`, we are able to overwrite React modules, but unfortunately there is not any good way of overwriting packager modules.  I'm trying to implement support for `unbundle` in `react-native-windows`, but unfortunately this file limits assets-based unbundles to the Android platform only.  This change will allow platforms other than Android to use unbundles, while still disabling assets unbundles for iOS.