Inherit WebSocket protocol scheme for http/https
When a WebSocket URL with protocol scheme http or https is used the
method getDefaultOrigin will fail to substitute it and just returns a
URL with empty protocol scheme leading to an opaque exception down the
line.

Thus, in case of http or https it should just inherit the protocol
scheme since the WebSocket is responsible for upgrading the HTTP/HTTPS
connection to the WebSocket protocol.