Correctly handle the case where HEADER_SEARCH_PATHS is a string
Fixes #12959

Would love some feedback on the test plan here. Modifying `project.pbxproj`
means that there's no test coverage for the (much more common) case where it's
actually an array. It might sense to add a new `pbxproj` fixture, but that's a
huge amount of overhead for such a small change.