Helping with helloworld on local machine
It is not just copy and paste into  index.ios.js or index.android.js to make the hello world app work on your local machine, as the text suggests. It requires the app's name to match the component which is registered with the registry. I did the Getting Started page before this and thus had the AwesomeProject app started and could paste into it, but naturally it did not work. To get it working, I originally just basically did find and replace HelloWorldApp -> AwesomeProject, which worked, but the better solution I found was in the stackoverflow topic I linked. If there is a better way to make it easier for beginners to test the hello world app in their local machine, I would be open to it.