Improve ReactModalHostView sizing logic
Setting the size of modal host's children according to the top level
FrameLayout we have, instead of querying screen size and manually
subtracting sizes of screen decorations. This way we don't need the
custom shadow node for the modal.

This also fixes a rare corner case that happens on certain Android
devices where the application windows are smaller than the actual
resolution of the screen and scaled up by the hardware scaler while
being composited.