Expose shake-to-show dev menu setting via the bridge
Summary: Apps commonly provide their own rage-shake menus or behaviors, including in dev builds where the dev menu is enabled on shake. Rather than try to override these settings via native code, it can be helpful to let the app define when to show the menu via the bridge.

Test Plan: Add the following code to `UIExplorerApp.ios.js` `componentDidMount`:

  NativeModules.DevSettings.setIsShakeToShowDevMenuEnabled(false);

And then rage shake the device (or simulator). Ensure that the menu no longer appears on shake.

Other tests:
- Ensure simulator Command-D shortcut still functions on simulator with setting enabled.
- Ensure calling `NativeModules.DevMenu.show()` still functions with setting enabled.
- Ensure calling the method with `false` and then `true` reverts the setting back to the original state.

Note that these settings are stored in user defaults, so setting this to `false` and then removing the code will not re-enable the menu; a subsequent call with `true` must occur.