Android: Make lineHeight accept decimal values
Make android-version accept a decimal
number as lineHeight.

According to the spec the property
line-height should accept decimal values
(and it does on iOS) but the android
version has the wrong data-type for the
shadowed method, resulting in a stacktrace
saying:
com.facebook.react.bridge.UnexpectedNativeTypeException: TypeError:
expected dynamic type `int64', but had type `double'

Setting it to a float makes it accept
decmial values as it should.

Credits where due, solution was given in this
issue: facebook/react-native#10607

Verification:
* Create an app without this fix:
* Create the same app with this fix:
- Leave line-height undefined. Behaviour
    is same as before.
- Set lineHeight to a integer number:
    Behaviour is same as before.
- Set lineHeight to a decimal number.
    Line height is now rendered with decimals.
* Run android integration tests to see nothing
  else broke.