Replace Navigator references to with existing components
Navigator page("docs/navigator.html") does not exist anymore b/c the
Navigator component itself has been removed from the package.

Replace Navigator references to other components that are less likely
to be changed over time(e.g. TextInput and Button components).

No plans to test as this change is on README of the repository.

No issue was created b/c this is not a bug on code nor a feature
request. And there already is an awareness of dead Navigation links
floating inside this repository(e.g. #13804).