Using some 3rd party libraries with React Native is not possible due to their reliance on some CommonJS/node module behavior not implemented by the React Native require polyfill from the packager.
This PR adds some of these behaviors:

Circular dependencies when exporting by assignment to the module.exports object. Should fix #3298 for good.
Exporting by assignment to the this object. Fixes #2862
The module.id represents a unique identifier that can be used in require() calls.
There were no tests for the require polyfill so I added one.