iOS: Avoid adding extra spaces to accessibility label
In some cases, RCTRecursiveAccessibilityLabel could return an accessibility label that had leading spaces, trailing spaces, or multiple spaces between words. This is because it always added a space before adding a label even if the label turned out to be empty.

This is fixed by being stricter about adding spaces.