Document Android native communication
Native communication is documented for iOS but not Android. Typically
pages in Guides (IOS) and Guides (Android) share a common
structure/content but provide language specific examples.

This pattern has been followed with the Android native communication
documentation. It is missing some content that the iOS documentation
contains, such as the layout computation flow. But provides a solid base
for future efforts.