Allow apps to manually control when the dev support tools are enabled
For multi-activity apps, apps that want fine grained control over when the dev support tools are enabled, or apps that frequently switch between native and react native, it isn't ideal to automatically enable and disable the dev tools in onHostPause and onHostResume. This allows apps to manually control when the dev tools are enabled.

Fixes #13135