Adds plugin install feature to `install` and `uninstall`
For react-native-windows (as well as other CLI plugins for react-native), the user should not be expected to know whether they need to use yarn or NPM to install a CLI plugin. We were running into issues where users installed react-native using yarn, but when following our tutorial for adding `rnpm-plugin-windows`, the CLI would break as the instructions currently say to use NPM. These extensions to `install` and `uninstall` will allow us to have a simple tutorial for how to install the `rnpm-plugin-windows`, as well as any other RN CLI plugin. The plugins use the existing CLI utilities to check if the project is yarn or NPM based, and then install the specified plugin accordingly.