Fix tvOS build
`TARGET_OS_IPHONE` is not the right way to differentiate between iOS and tvOS. From a bug report to Apple (http://www.openradar.me/22639221) about `TARGET_OS_IPHONE` being true on tvOS:

"This is not a bug. tvOS is an evolutionary subclass of iOS (formerly "iPhone OS"). You need to use TARGET_OS_IOS instead, which is NOT defined to 1 on tvOS."