iOS: Fix RCTDeviceInfo Dimensions for iPad multitasking
**Issue:**
On iOS Dimensions `window` and `screen` both get resolved to the UIScreen bounds. This will always give us the dimensions in terms of the current physical screen currently in use. For the `window` value of Dimensions this sh    ould be considered incorrect. With the introduction of multitasking on iOS we now have the ability to have our app rendered on only part of the screen and the `window` value should reflect the dimensions our app is currently rendering in.

**Example:**
When run on iPad this gist will render a red view with a black border using `Dimensions.get('window').width`. When rotating, this view will display correctly. If you instead run this gist while in multitasking mode, the view will render incorrectly as `Dimensions.get('window').width` is still returning the full width of the physical screen.
https://gist.github.com/kipricker/c87b606ba439e7b09fd29303bf324065

**Proposed fix:**
For updating, instead of using the rotation of the screen to determine an update in our `window` dimensions, we can have RCTWindowManager report our `availableSize` when it changes. This will update us properly on window size changes including size changes on rotation.

For values, `screen` will still use the bounds of UIScreen, and `window` will now use the bounds of the current rootViewController.view.

**Test Plan:**
To test you can watch the values of the `didUpdateDimensions` DeviceEvent. The following should be true:
Dimensions properly update on rotation.
Dimensions properly update in multitasking mode when app window resizes.
Dimensions only update once with the same result when doing the above actions.
Dimensions only update when the app is in the foreground.