Account for header, footer, and content container padding in VirtualizedList
VirtualizedList leans heavily upon frame estimates to power its
heuristics. Despite explicit support for headers and footers (and
implicit support for top and bottom padding via `contentContainerStyle`),
the estimates fail to account for the heights of these additions to the
top and bottom of the scrollable content area.

This commit addresses some of the inconsistencies, though I'm certain
that I missed some calculations. Notably, it fails to address horizontal
VirtualizedLists, and it doesn't address any of the sticky header logic,
which may or may not be wrong (it was too difficult to tell).