Android: Fix sizing of TextInput when font scale != 1
Sometimes, TextInputs on Android have the wrong size. Specifically, this occurs when the user has changed the system's font scale setting, the TextInput has `allowFontScaling={false}`, and the `fontSize` prop isn't set.

To fix this, TextInput's measure function needs to take into account the value of `allowFontScaling`.

**Test Plan**

Created a test app with a multiline `TextInput` which had `numberOfLines` set to `3`. Changed the system's font scale setting and verified that the `TextInput` was tall enough to show exactly 3 lines when `allowFontScaling` was both `true` and `false`.

Adam Comella
Microsoft Corp.