add programmatic variant for permission check
It can happen that location access is denied by the user because of them checking the "don't show again" checkbox in the permission dialog. When this is the case, every subsequent request of accessing the location will make the app crash, even when the permission is added to the manifest.