Relax error conditions for InitializeCore
This patch addresses an issue with non-native platforms using `expo` and `react-native-windows` with React Native 0.49 and 0.50, and conveniently also ensures `react-native-web` does not error either, if that has not been updated.

Summary:

Alternative platform implementations of React Native may not implement the modules `PlatformConstants` or `ReactNativeVersion`. This adds appropriate warnings when these modules are not implemented.

There was some discussion on #16300 about whether a version mismatch should throw an error. This patch does not change the version mismatch behavior and retains throwing an error. The lines above and below `throw new Error` are kept the same, so there will hopefully not be a merge conflict if that changes.

The term "Native version" was confusing to me, but I left it as is. I would suggest "Platform version", since the term "Native" is overloaded a bit here? Again, this commit is agnostic about that terminology.